1. permutation matters 2. count every operation time 3. lubm10240-l7 is most important, some tests will run on lubm320-l7

+++ selfTests/OperationsTiming.cpp
@@ -14,8 +14,8 @@ using namespace combblas;
 template<class NT>
 class PSpMat {
 public:
-    typedef SpDCCols<int, NT> DCCols;
-    typedef SpParMat<int, NT, DCCols> MPI_DCCols;
+    typedef SpDCCols<int64_t, NT> DCCols;
+    typedef SpParMat<int64_t, NT, DCCols> MPI_DCCols;
 };
 
 #define ElementType int
